Safari-609.1.11.1/LayoutTests/fast/inline/inline-fixed-position-boundingbox-expected.txt

Bug 91451: REGRESSION: RenderInline::absoluteQuads produces incorrect results for fixed position.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS inlineRect.left is parentRect.left
